fix: Added Main cmp and fix autoconnect
