+++ src/components/Main/index.ts
@@ -1,0 +1,2 @@
+export { default } from './cmp'
+export type { MainProps } from './types'
