+++ src/components/Main/types.ts
@@ -1,0 +1,5 @@
+import { ReactNode } from 'react'
+
+export type MainProps = {
+  children: ReactNode
+}
